+++ lizard_languages/rust.py
@@ -1,5 +1,5 @@
 '''
-Language parser for Go lang
+Language parser for Rust lang
 '''
 
 from .code_reader import CodeReader
@@ -16,8 +16,13 @@ class RustReader(CodeReader, CCppCommentsMixin):
                       'case', 'match', 'where'])
 
     def __init__(self, context):
-        super(RustReader, self).__init__(context)
+        super().__init__(context)
         self.parallel_states = [RustStates(context)]
+
+    @staticmethod
+    def generate_tokens(source_code, addition='', token_class=None):
+        addition = r"|(?:'\w+\b)"  # lifetimes, labels
+        return CodeReader.generate_tokens(source_code, addition, token_class)
 
 
 class RustStates(GoLikeStates):  # pylint: disable=R0903