+++ test/test_languages/testGo.py
@@ -17,7 +17,7 @@ class Test_parser_for_Go(unittest.TestCase):
     def test_no_function(self):
         result = get_go_function_list('''
         for name, ok := range names; ok {
-                print("Hello, \(name)!")
+                print("Hello, \\(name)!")
             }
                 ''')
         self.assertEqual(0, len(result))
@@ -68,6 +68,44 @@ class Test_parser_for_Go(unittest.TestCase):
                 ''')
         self.assertEqual(2, result[0].cyclomatic_complexity)
 
+    def test_one_function_with_return_empty_interface(self):
+        result = get_go_function_list('''
+            func sayGoodbye() interface{} {
+                if ++diceRoll == 7 { diceRoll = 1 }
+            }
+                ''')
+        self.assertEqual(1, len(result))
+        self.assertEqual("sayGoodbye", result[0].name)
+        self.assertEqual(3, result[0].length)
+
+    def test_nest_function(self):
+        result = get_go_function_list('''
+            func sayGoodbye() {
+                f1 := func() {}
+                f2 := func(n int) {}
+                f3 := func() int {
+                    return 0
+                }
+            }
+                ''')
+        self.assertEqual(4, len(result))
+
+        self.assertEqual("", result[0].name)
+        self.assertEqual("", result[0].long_name)
+        self.assertEqual(1, result[0].length)
+
+        self.assertEqual("", result[1].name)
+        self.assertEqual(" n int", result[1].long_name)
+        self.assertEqual(1, result[1].length)
+        self.assertEqual(['n int'], result[1].full_parameters)
+
+        self.assertEqual("", result[2].name)
+        self.assertEqual("", result[2].long_name)
+        self.assertEqual(3, result[2].length)
+
+        self.assertEqual("sayGoodbye", result[3].name)
+        self.assertEqual(7, result[3].length)
+
     def test_interface(self):
         result = get_go_function_list('''
 			type geometry interface{
@@ -89,3 +127,42 @@ class Test_parser_for_Go(unittest.TestCase):
                 ''')
         self.assertEqual(0, len(result))
 
+    def test_struct_with_func_followed_by_function_with_receiver(self):
+        result = get_go_function_list('''
+            type Geometry struct {
+                isEqual func(float64, float64) error
+            }
+
+            func (g *Geometry) sayGoodbye() { }
+                ''')
+
+        self.assertEqual(1, len(result))
+        self.assertEqual("sayGoodbye", result[0].name)
+
+    def test_interface_with_func_followed_by_function_with_receiver(self):
+        result = get_go_function_list('''
+            type MyComparator struct{}
+
+            type Comparator interface {
+                Handle(func(int) string)
+            }
+
+            func (m MyComparator) Handle(f func(int) string) {}
+                ''')
+
+        self.assertEqual(1, len(result))
+        self.assertEqual("Handle", result[0].name)
+
+    def test_sql_query_with_question_marks(self):
+        result = get_go_function_list('''
+            func getQuery(dbIndex uint32, tbIndex uint32) string {
+                query := fmt.Sprintf(`INSERT INTO online_docs_%d.online_docs_notify_%d
+                (a, b, c, d, e, f, g, h, i, j)
+                VALUES (?, ?, ?, ?, ?, ?, ?, FROM_UNIXTIME(?), ?, %d)`,
+                dbIndex, tbIndex, notifyStatusNew)
+                return query
+            }
+                ''')
+        self.assertEqual(1, len(result))
+        self.assertEqual("getQuery", result[0].name)
+        self.assertEqual(1, result[0].cyclomatic_complexity)